Moved SelectComponent to profile components folder
